Fix Windows build.

+++ src/libs/filemanager/fileinfodialog.cpp
@@ -91,11 +91,11 @@ static QString sizeToString(qint64 size)
     if (size <= 0)
         return FileInfoDialog::tr("0 b");
 
-    double power = log(size)/log(1024.0);
+    double power = log((double)size)/log(1024.0);
     int intPower = (int)power;
     intPower = intPower >= 8 ? 8 - 1 : intPower;
 
-    double normSize = size / pow(1024, intPower);
+    double normSize = size / pow(1024.0, intPower);
     return FileInfoDialog::tr("%1 %2").arg(normSize, 0, 'f', intPower > 0 ? 2 : 0).arg(strings[intPower]);
 }
